# [Model][Perf] Improve sage sampling performance