--- a/tests/regression/README.md
+++ b/tests/regression/README.md
@@ -20,6 +20,7 @@ The basic use is execute a script, and get the needed results out of the printed
 
 ```bash
 docker run --name dgl-reg --rm --hostname=reg-machine --runtime=nvidia -dit dgllib/dgl-ci-gpu:conda /bin/bash
+docker cp /home/ubuntu/asv_data dgl-reg:/root/asv_data/
 docker exec dgl-reg bash /root/asv_data/run.sh
 docker cp dgl-reg:/root/regression/dgl/asv/. /home/ubuntu/asv_data/  # Change /home/ubuntu/asv to the path you want to put the result
 docker stop dgl-reg