--- a/tests/regression/run.sh
+++ b/tests/regression/run.sh
@@ -1,10 +1,22 @@
+#!/bin/bash
+set -e
+
+if [ $# -ne 2 ]; then
+    echo "run.sh <repo> <branch>"
+    exit 1
+fi
+
+REPO=$1
+BRANCH=$2
+
 . /opt/conda/etc/profile.d/conda.sh
 
 cd ~
 mkdir regression
 cd regression
 # git config core.filemode false
-git clone --recursive https://github.com/dmlc/dgl.git 
+git clone --recursive https://github.com/$REPO/dgl.git 
+git checkout $BRANCH
 cd dgl
 mkdir asv
 cp -r ~/asv_data/* asv/