… examples/pytorch/graphsage/train_full.py