## Description
<!-- Brief description. Refer to the related issues if existed.
It'll be great if relevant reviewers can be assigned as well.-->

* Turn on OMP in `rowwise_pick.h`
* Remove the costly dropout in the input layer.
* Make dedicated scripts for single process and multi-GPU.

Benchmarks:
* On p3.2xlarge, with `--num-workers=4`, each epoch in DGL takes 21.7179 seconds while PyG takes 33.4735 seconds (1.5x)
* On m5d.24xlarge, with `--num-workers=8`, each epoch in DGL takes 22.5483 seconds while PyG takes 24.3307 seconds (1.07x)

@VoVAllen Tried to follow your instruction to add regression test but failed. WE NEED BETTER INSTRUCTIONS! Several issues:
* Many hard-coded paths.
* The script pulls the repo from dmlc/dgl but when we run it in local, the changes are usually in our fork, or simply just in local.
* Since rebuilding everything is quite slow, consider provide a shortcut to allow use our locally built binaries to generate reports.

## Checklist
Please feel free to remove inapplicable items for your PR.
- [x] The PR title starts with [$CATEGORY] (such as [NN], [Model], [Doc], [Feature]])
- [x] Changes are complete (i.e. I finished coding on this PR)
- [x] All changes have test coverage
- [x] Code is well-documented
- [x] To the my best knowledge, examples are either not affected by this change,
      or have been fixed to be compatible with this change
- [x] Related issue is referred in this PR